+++ tests/strategies/test_single_device_strategy.py
@@ -11,9 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import pickle
+
 import torch
 
 from pytorch_lightning import Trainer
+from pytorch_lightning.core.optimizer import LightningOptimizer
 from pytorch_lightning.strategies import SingleDeviceStrategy
 from tests.helpers.boring_model import BoringModel
 from tests.helpers.runif import RunIf
@@ -53,3 +56,23 @@ def test_single_gpu():
     assert model.device == torch.device("cpu")
     cuda_memory = torch.cuda.memory_allocated()
     assert cuda_memory < model.start_cuda_memory
+
+
+class MockOptimizer:
+    ...
+
+
+def test_strategy_pickle():
+    strategy = SingleDeviceStrategy("cpu")
+    optimizer = MockOptimizer()
+
+    strategy.optimizers = [optimizer]
+    assert isinstance(strategy.optimizers[0], MockOptimizer)
+    assert isinstance(strategy._lightning_optimizers[0], LightningOptimizer)
+
+    state = pickle.dumps(strategy)
+    # dumping did not get rid of the lightning optimizers
+    assert isinstance(strategy._lightning_optimizers[0], LightningOptimizer)
+    strategy_reloaded = pickle.loads(state)
+    # loading restores the lightning optimizers
+    assert isinstance(strategy_reloaded._lightning_optimizers[0], LightningOptimizer)